Use powf instead of pow in sigmoid

+++ src/iop/sigmoid.c
@@ -151,8 +151,8 @@ static inline float generalized_loglogistic_sigmoid(const float value, const flo
   // The following equation can be derived as a model for film + paper but it has a pole at 0
   // magnitude * powf(1.0 + paper_exp * powf(film_fog + value, -film_power), -paper_power);
   // Rewritten on a stable form including a check for negative values:
-  const float film_response = film_fog + value > 0.0f ? pow(film_fog + value, film_power) : 0.0f;
-  return magnitude * pow(film_response / (paper_exp + film_response), paper_power);
+  const float film_response = film_fog + value > 0.0f ? powf(film_fog + value, film_power) : 0.0f;
+  return magnitude * powf(film_response / (paper_exp + film_response), paper_power);
 }
 
 void commit_params(dt_iop_module_t *self, dt_iop_params_t *p1, dt_dev_pixelpipe_t *pipe, dt_dev_pixelpipe_iop_t *piece)
